backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
easyocr==1.7.1
opencv-python==4.9.0.80
numpy==1.24.3
pillow==10.2.0
rapidfuzz==3.6.1
pydantic==2.5.3
